test: use structuredClone polyfill in tests as well
