+++ packages/vitest-pool-workers/test/global-setup.ts
@@ -4,6 +4,7 @@ import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
 import type { GlobalSetupContext } from "vitest/node";
+import "../src/shared/structured-clone-polyfill";
 
 function packPackage(packDestinationPath: string, packagePath: string) {
 	const output = childProcess.execSync(
